flask_task/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
